Update: implementing the calculator for calculating caffeine decays.

+++ src/caffine_info/caffeine_info.rs
@@ -4,6 +4,6 @@ use serde::{Deserialize, Serialize};
 pub struct CaffeineInfo {
     pub(crate) caffeine_mg: f64,
     pub(crate) time: f64,
-    pub(crate) method: u64,
+    pub(crate) method: u8,
     pub(crate) bottle_ml: f64,
 }
